refactor(usage): applied the new folder structure to the usage section related components
