+++ src/features/usage/components/CategoryFormat/IndeterminateCheckbox.jsx
